Move fixture tests to new, dedicated module `test_fixtures.py`.
