+++ tests/test_job.py
@@ -985,17 +985,6 @@ class TestJob(RQTestCase):
 
         assert job_C.result
 
-    def test_rpush_fixture(self):
-        fixtures.rpush('foo', 'bar')
-        assert self.testconn.lrange('foo', 0, 0)[0].decode() == 'bar'
-
-    def test_start_worker_fixture(self):
-        queue = Queue(name='testing', connection=self.testconn)
-        queue.enqueue(fixtures.say_hello)
-        conn_kwargs = self.testconn.connection_pool.connection_kwargs
-        fixtures.start_worker(queue.name, conn_kwargs, 'w1', True)
-        assert not queue.jobs
-
     def test_execution_order_with_sole_dependency(self):
         queue = Queue(connection=self.testconn)
         key = 'test_job:job_order'
